requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
avro-python3==1.9.1
certifi==2018.11.29
cffi==1.13.2
chardet==3.0.4
confluent-kafka==1.3.0
cycler==0.10.0
Cython==0.29.6
fastavro==0.22.9
future==0.18.2
idna==2.8
isort==4.3.15
kafka-python==1.4.7
kiwisolver==1.0.1
labelImg==1.8.3
lazy-object-proxy==1.3.1
lxml==4.3.3
matplotlib==3.0.3
mccabe==0.6.1
multidict==4.7.3
mysql-connector==2.2.9
mysql-connector-python==8.0.18
numpy==1.16.2
oauthlib==3.1.0
object-detection==0.1
Pillow==5.4.1
pipenv==2018.11.26
protobuf==3.1.0
py-mysql==1.0
py4j==0.10.7
pycares==3.1.0
pycparser==2.19
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.3.1
PyQt5==5.12.2
PyQt5-sip==4.19.17
PySocks==1.7.1
pyspark==2.4.4
python-dateutil==2.8.0
python-twitter==3.5
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
slackclient==2.5.0
tensorflow==0.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tweepy==3.8.0
typed-ast==1.3.1
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
wrapt==1.11.1
yarl==1.4.2
